Remove DObject::FreeIndices and DObject::Objects.
These were a remnant from before the introduction of szp, used to keep
a pointer to all existing DObjects so that they could be iterated over
to null any pointers to deleted DObjects. Now they are useless, as
well as buggy due to RemoveFromArray being called twice sometimes
if DObject::Destroy is called.